Add test path that hits a more traditional local test database instead of Docker
Here, add a testing path that hits a more traditional local test
database URL instead of requiring that the tests go through Docker, and
making the suite a little more like other River projects and able to
share a test database with them (`river_test` being a common name
convention between them all).

I've also changed the test code to rollback test transactions after each
test case so as to not leave any artifacts in the database that might
affect other test suites.

I've left the Docker path in as well, and it can be activated by sending
the `RIVER_USE_DOCKER` env var when running tests:

    RIVER_USE_DOCKER=true rye test

I also added code so that the tests get type checked as well as the main
code, which I found helped detect bugs, and being a Python beginner,
help better understand what exactly various fixture functions were
supposed to be returning. I didn't type everything, mostly just common
fixture functions. To make this possible I had to make a few changes
like having the main package declare itself as typed by exporting a
`py.typed` file, and also making `tests` a module with `__init__.py` so
MyPy could resolve common utilities in it like `SimpleArgs`.
