+++ mypy.ini
@@ -1,4 +1,8 @@
 [mypy]
 
-[mypy-riverqueue.sqlcdb]
-ignore_errors = True
+# testcontainers is typed, but it doesn't correctly declare itself as such.
+# Hopefully it can be fixed one day:
+#
+# https://github.com/testcontainers/testcontainers-python/issues/305
+[mypy-testcontainers.*]
+ignore_missing_imports = True
