+++ tests/client_test.py
@@ -1,5 +1,3 @@
-import json
-from dataclasses import dataclass
 from datetime import datetime, timezone
 from unittest.mock import MagicMock, patch
 
@@ -7,6 +5,8 @@ import pytest
 
 from riverqueue.client import Client
 from riverqueue.models import InsertOpts, UniqueOpts
+
+from tests.simple_args import SimpleArgs
 
 
 @pytest.fixture
@@ -17,15 +17,6 @@ def mock_driver():
 @pytest.fixture
 def client(mock_driver):
     return Client(mock_driver)
-
-
-@dataclass
-class SimpleArgs:
-    kind: str = "simple"
-
-    @staticmethod
-    def to_json() -> str:
-        return json.dumps({"job_num": 1})
 
 
 @patch("datetime.datetime")
